requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database Connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Apache Airflow (for workflow orchestration)
apache-airflow>=2.8.0

# Data Validation
great-expectations>=0.18.0

# File Format Support
pyarrow>=14.0.0
openpyxl>=3.1.0

# API and Web
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
